requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
awscli==1.22.34
billiard==4.2.0
blinker==1.4
botocore==1.23.34
celery==5.4.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
cron-descriptor==1.4.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0.6
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-filter==24.2
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.17.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jmespath==0.10.0
keyring==23.5.0
kombu==5.3.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
Pillow==9.0.1
prompt_toolkit==3.0.47
pyasn1==0.4.8
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==5.4.1
redis==5.0.7
requests==2.25.1
rest-framework-simplejwt==0.0.2
roman==3.3
rsa==4.8
s3transfer==0.5.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.0
systemd-python==234
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
zipp==1.0.0
